+++ resources/views/TeamAssignment/newIndex.blade.php
@@ -113,14 +113,11 @@
                                                 @if(count($compliance)>0)
                                                     <div class="card-body text-white">
                                                         @foreach($compliance as $compliance)
-                                                            <div class="card bg-success mb-2 text-center" id="dismissTeam">
+                                                            <div class="card bg-success mb-2 text-center">
                                                                 <div style="margin-top: 10px;">
                                                                     <a href="#" onclick="selectTugboatTeam({{$compliance->intTAssignID}})"class="teamTugboat text-white">
                                                                         <div class="card-header">
                                                                             {{$compliance->strName}}
-                                                                            <div class="float-right">
-                                                                                <a data-dismiss="#dismissTeam" class="btn btn-icon"><i class="ion ion-close"></i></a>
-                                                                            </div>
                                                                         </div>
                                                                     </a>
                                                                 </div>
@@ -129,8 +126,8 @@
                                                     </div>
                                                 @else
                                                     <div class="card-body">
-                                                        <div class=" text-danger text-center">
-                                                            <i class="fas fa-times mr-2"></i> No Available Tugboats <i class="fas fa-times ml-2"></i>
+                                                        <div class="alert alert-danger text-center">
+                                                            <i class="fas fa-exclamation-triangle mr-2"></i>NO AVAIABLE OWNED TUGBOATS FOUND!
                                                         </div>
                                                     </div>
                                                 @endif
@@ -149,14 +146,11 @@
                                                 @if(count($tugboatAvail)>0)
                                                     <div class="card-body text-white">
                                                         @foreach($tugboatAvail as $tugboatAvail)
-                                                            <div class="card bg-info mb-2 text-center" id="dismissTeam">
+                                                            <div class="card bg-info mb-2 text-center">
                                                                 <div style="margin-top: 10px;">
                                                                     <a href="#" onclick="selectTugboatTeam({{$tugboatAvail->intTAssignID}})"class="teamTugboat text-white">
                                                                         <div class="card-header">
                                                                             {{$tugboatAvail->strName}}
-                                                                            <div class="float-right">
-                                                                                <a data-dismiss="#dismissTeam" class="btn btn-icon"><i class="ion ion-close"></i></a>
-                                                                            </div>
                                                                         </div>
                                                                     </a>
                                                                 </div>
@@ -165,8 +159,8 @@
                                                     </div>
                                                 @else
                                                     <div class="card-body">
-                                                        <div class=" text-danger text-center">
-                                                            <i class="fas fa-times mr-2"></i> No Occupied Tugboats <i class="fas fa-times ml-2"></i>
+                                                        <div class="alert alert-danger text-center">
+                                                            <i class="fas fa-exclamation-triangle mr-2"></i>NO OWNED OCCUPIED TUGBOATS FOUND!
                                                         </div>
                                                     </div>
                                                 @endif
@@ -177,7 +171,7 @@
                                                         <div class="card bg-info" style="margin-top: 0px;">                                      
                                                             @foreach($tugboatAvail as $tugboatAvail)
                                                                 <a href="#" onclick="selectTugboatTeam({{$tugboatAvail->intTAssignID}})"class="teamTugboat text-white">
-                                                                    <div class="card-header" id="dismissTeam">
+                                                                    <div class="card-header">
                                                                         {{$tugboatAvail->strName}}
                                                                         <div class="float-right">
                                                                             <a data-dismiss="#dismissTeam" class="btn btn-icon"><i class="ion ion-close"></i></a>
@@ -205,9 +199,9 @@
                                             </div>
                                             <div class="collapse show" id="ownedUnderRepairT">
                                                 <div class="card-body">
-                                                    <div class=" text-danger text-center">
-                                                        <i class="fas fa-times mr-2"></i> No Disabled Tugboat <i class="fas fa-times ml-2"></i>
-                                                    </div>
+                                                        <div class="alert alert-danger text-center">
+                                                            <i class="fas fa-exclamation-triangle mr-2"></i>NO DISABLED TUGBOATS FOUND!
+                                                        </div>
                                                 </div>
                                             </div>
                                         </div>
@@ -254,8 +248,8 @@
                                                     </div>
                                                     <div class="collapse show" id="receivedOccupiedT">
                                                         <div class="card-body">
-                                                            <div class=" text-danger text-center">
-                                                                <i class="fas fa-times mr-2"></i> No Occupied Tugboat <i class="fas fa-times ml-2"></i>
+                                                            <div class="alert alert-danger text-center">
+                                                                <i class="fas fa-exclamation-triangle mr-2"></i>NO RECEIVED OCCOUPIED TUGBOATS FOUND!
                                                             </div>
                                                         </div>
                                                     </div>