+++ resources/views/Templates/newSidebar.blade.php
@@ -74,6 +74,7 @@
                                 </li>
                                 <li id="tDispatch"><a href="#" class="has-dropdown waves-effect"><i class="ion ion-android-boat"></i>Dispatch &amp; Hauling</a>
                                     <ul class="menu-dropdown">
+                                        <li id="menuForwardReq"><a href="/administrator/transactions/forwardrequests"><i class="ion-ios-upload"></i>Forward Requests</a></li>
                                         <li id="menuJobOrder"><a class="waves-effect" href="/administrator/transactions/joborders"><i class="ion ion-ios-list"></i>Job Order</a></li>
                                         <li id="menuTugboatAssignment"><a class="waves-effect" href="/administrator/transactions/tugboatassignment"><i class="ion ion-android-boat"></i>Tugboat Assignment</a></li>
                                         <li id="menuTeamBuilder"><a class="waves-effect" href="/administrator/transactions/teamassignment"><i class="ion ion-ios-people"></i>Team Builder</a></li>
@@ -96,9 +97,9 @@
                             </ul>
                         @elseif(Auth::user()->enumUserType == 'Affiliates')
                             <ul class="menu-dropdown">
-                                
                                 <li><a href="#" class="has-dropdown waves-effect"><i class="ion ion-android-boat"></i>Dispatch &amp; Hauling</a>
                                     <ul class="menu-dropdown">
+                                        <li id="menuForwardReq"><a href="/administrator/transactions/forwardrequests"><i class="ion-ios-upload"></i>Forward Requests</a></li>
                                         <li id="menuJobOrder"><a class="waves-effect" href="/administrator/transactions/joborders"><i class="ion ion-ios-list"></i>Job Order</a></li>
                                         <li id="menuTugboatAssignment"><a class="waves-effect" href="/administrator/transactions/tugboatassignment"><i class="ion ion-android-boat"></i>Tugboat Assignment</a></li>
                                         <li id="menuTeamBuilder"><a class="waves-effect" href="/administrator/transactions/teamassignment"><i class="ion ion-ios-people"></i>Team Builder</a></li>